move gradient ascent to utils, adapt imports.
