+++ sbi/analysis/__init__.py
@@ -4,6 +4,5 @@ from sbi.analysis.conditional_density import (
     conditonal_potential,
     eval_conditional_density,
 )
-from sbi.analysis.gradient_ascent import gradient_ascent
 from sbi.analysis.plot import conditional_pairplot, pairplot
 from sbi.analysis.sensitivity_analysis import ActiveSubspace
